+++ src/pkcs15init/pkcs15-westcos.c
@@ -249,7 +249,7 @@ static int westcos_pkcs15init_generate_key(sc_profile_t *profile,
 
 	if (BN_set_word(bn, RSA_F4) != 1 ||
 	    EVP_PKEY_keygen_init(pctx) != 1 ||
-	    EVP_PKEY_CTX_set_rsa_keygen_bits(pctx, key_info->modulus_length) != 1 ||
+	    EVP_PKEY_CTX_set_rsa_keygen_bits(pctx, (int)key_info->modulus_length) != 1 ||
 #if OPENSSL_VERSION_NUMBER >= 0x30000000L
 	    EVP_PKEY_CTX_set1_rsa_keygen_pubexp(pctx, bn) != 1) {
 #else
@@ -319,7 +319,7 @@ static int westcos_pkcs15init_generate_key(sc_profile_t *profile,
 	r = sc_pkcs15init_create_file(profile, p15card, prkf);
 	if(r) goto out;
 
-	r = sc_pkcs15init_update_file(profile, p15card, prkf, p, lg);
+	r = sc_pkcs15init_update_file(profile, p15card, prkf, p, (unsigned int)lg);
 	if(r) goto out;
 
 out:
